fix(tri): drop/restore supplier FKs around upstream 0115 INT widen
Upstream 0115 changes supplier.id from TINYINT to INT. TRI tables
still referenced it, so the ALTER failed. Drop those FKs first,
widen tri_* supplier_id, then put the constraints back.

+++ db/migrations/9000_tri_client_tags.sql
@@ -2,7 +2,7 @@
 
 CREATE TABLE IF NOT EXISTS tri_tags (
   id           BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
-  supplier_id  TINYINT UNSIGNED NOT NULL,
+  supplier_id  INT UNSIGNED NOT NULL,
   name         VARCHAR(80) NOT NULL,
   slug         VARCHAR(80) NOT NULL,
   color        VARCHAR(7) NOT NULL DEFAULT '#6366f1',
